A dataset of criticality benchmarks calculations

A dataset of criticality benchmarks calculations

data(sendis)

Format

A data frame with 12869 observations of 25 variables:

FULLID

Full ICSBEP identifier of benchmark evaluation, case and subcase, e.g.: PU-MET-FAST-001-001

SHORTID

Abbreviated ICESBEP identifier of benchmark evaluation, case and subcase, e.g.: PMF-1-1

MODEL

Model description in case different models exist for same benchmark

CASETYPE

Type or family of benchmark case, e.g. : PU-MET-FAST

FISS

ICSBEP identifier of fissile material in benchmark, e.g.: HEU, LEU, U233

FORM

ICSBEP identifier of material form in benchmark, e.g.: MET, SOL, COMP

SPEC

ICSBEP identifier of neutron spectrum category in benchmark, e.g.: THERM, INTER, FAST

EXPVAL

Experimental value of benchmark Keff, as reported by DICE

EXPERR

Uncertainty associated to the experimental value of benchmark Keff, as reported by DICE

EALF

Energy corresponding to the average neutron lethargy causing fission

AFGE

Average neutron energy causing fission

CODE

Code used for the calculation, can include version number

INST

Name of institute (can also be used to include author of calculations) - Important : this field is also used to distinguish different benchmarking suites

LIBVER

Library used in calculations, including version number, e.g.: JEFF-3.3

LIB

Library used in calculations, e.g.: JEFF, ENDFB

VER

Version of library used in calculations, e.g.: 3.1.1, 3.3, 7.1, 8.0

CALCVAL

Calculated value of Keff

CALCERR

Uncertainty of calculated value (statistical for Monte Carlo)

NCASES

Number of cases in a benchmarking suite

COVERE

CALCVAL / EXPVAL

TOTERR

sqrt(CALCERR^2 + EXPERR^2)

RESIDUAL

(CALCVAL - EXPVAL)/TOTERR

CUMUL

Used for cumulative chi-square plots

CHISQ

Chi-square value for benchmark suite for a specific INST, e.g. for INST=="NEA", NCASES==123

Source

www.sendis.org

Examples

data(sendis)
head(sendis)

#>                   FULLID SHORTID MODEL       CASETYPE FISS FORM  SPEC
#> 1  HEU-COMP-FAST-003-001 HCF-3-1  Only  HEU-COMP-FAST  HEU COMP  FAST
#> 2  HEU-COMP-FAST-003-001 HCF-3-1  Only  HEU-COMP-FAST  HEU COMP  FAST
#> 3  HEU-COMP-FAST-003-001 HCF-3-1  Only  HEU-COMP-FAST  HEU COMP  FAST
#> 4  HEU-COMP-FAST-003-001 HCF-3-1  Only  HEU-COMP-FAST  HEU COMP  FAST
#> 5  HEU-COMP-FAST-003-001 HCF-3-1  Only  HEU-COMP-FAST  HEU COMP  FAST
#> 6 HEU-COMP-INTER-003-001 HCI-3-1  Only HEU-COMP-INTER  HEU COMP INTER
#>          LIBVER   LIB     VER       CODE INST NCASES  EALF  AFGE EXPVAL EXPERR
#> 1    JEFF-3.1.1  JEFF   3.1.1  MCNP6.1.1  NRG   1904 23700 20100      1 0.0085
#> 2     JENDL-4.0 JENDL     4.0  MCNP6.1.1  NRG   1904 23700 20100      1 0.0085
#> 3      JEFF-3.3  JEFF     3.3 MCNP.6.1.1  NRG   2139 23700 20100      1 0.0085
#> 4   ENDFB-VII.1 ENDFB   VII.1  MCNP6.1.1  NRG   1904 23700 20100      1 0.0085
#> 5 ENDFB-VIII.b4 ENDFB VIII.b4 MCNP.6.1.1  NRG   2139 23700 20100      1 0.0085
#> 6   ENDFB-VII.1 ENDFB   VII.1      MCNP6 IAEA    434  3190  7070      1 0.0057
#>   CALCVAL CALCERR  COVERE      TOTERR  RESIDUAL CUMUL CHISQ
#> 1 0.98229 0.00042 0.98229 0.008510370 -2.080991     0  5.32
#> 2 0.98982 0.00046 0.98982 0.008512438 -1.195897     0  8.68
#> 3 0.98466 0.00041 0.98466 0.008509882 -1.802610     0  5.33
#> 4 0.98672 0.00038 0.98672 0.008508490 -1.560794     0  5.49
#> 5 0.98667 0.00040 0.98667 0.008509407 -1.566502     0  4.73
#> 6 1.00489 0.00011 1.00489 0.005701061  0.857735     0  5.74